Share conversation handler base
